+++ data-protocols/ids/ids-api-multipart-dispatcher-v1/src/test/java/org/eclipse/dataspaceconnector/ids/api/multipart/client/MultipartDispatcherIntegrationTest.java
@@ -197,7 +197,7 @@ class MultipartDispatcherIntegrationTest extends AbstractMultipartDispatcherInte
     void testSendContractAgreementMessage() throws Exception {
         var contractAgreement = ContractAgreement.Builder.newInstance()
                 .id("1:23456").consumerAgentId("consumer").providerAgentId("provider")
-                .policy(Policy.Builder.newInstance().build())
+                .policyId("policyId")
                 .assetId(UUID.randomUUID().toString())
                 .build();
 
@@ -212,6 +212,7 @@ class MultipartDispatcherIntegrationTest extends AbstractMultipartDispatcherInte
                 .protocol(Protocols.IDS_MULTIPART)
                 .contractAgreement(contractAgreement)
                 .correlationId("1")
+                .policy(Policy.Builder.newInstance().id("policyId").build())
                 .build();
 
         var result = multipartDispatcher.send(MultipartMessageProcessedResponse.class, request, () -> null).get();