+++ extensions/azure/data-plane/storage/src/test/java/org/eclipse/dataspaceconnector/azure/dataplane/azurestorage/pipeline/AzureDataSourceToDataSinkTest.java
@@ -35,6 +35,8 @@ import static java.lang.String.format;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -44,6 +46,7 @@ class AzureDataSourceToDataSinkTest {
     Monitor monitor = mock(Monitor.class);
     FakeBlobAdapter fakeSource = new FakeBlobAdapter();
     FakeBlobAdapter fakeSink = new FakeBlobAdapter();
+    FakeBlobAdapter fakeCompletionMarker = new FakeBlobAdapter();
     String sourceAccountName = AzureStorageTestFixtures.createAccountName();
     String sourceContainerName = AzureStorageTestFixtures.createContainerName();
     String sourceSharedKey = AzureStorageTestFixtures.createSharedKey();
@@ -84,6 +87,12 @@ class AzureDataSourceToDataSinkTest {
                 fakeSource.name,
                 sinkSharedKey
         )).thenReturn(fakeSink);
+        when(fakeSinkFactory.getBlobAdapter(
+                eq(sinkAccountName),
+                eq(sinkContainerName),
+                argThat(name -> name.endsWith(".complete")),
+                eq(sinkSharedKey)
+        )).thenReturn(fakeCompletionMarker);
 
         var dataSink = AzureStorageDataSink.Builder.newInstance()
                 .accountName(sinkAccountName)